Split session and repository
